+++ public/video-ui/src/routes.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Router, Route, browserHistory, IndexRedirect, Redirect } from 'react-router';
 
-import Videos from './components/Videos/Videos';
+import Search from './pages/Search';
 import VideoDisplay from './components/Video/Display';
 import VideoUpload from './components/VideoUpload/VideoUpload';
 import Help from './pages/Help';
@@ -12,7 +12,7 @@ export const routes = (
   <Router history={browserHistory}>
     <Route path="/" component={ReactApp}>
       <IndexRedirect to="/videos" />
-      <Route path="/videos" component={Videos} />
+      <Route path="/videos" component={Search} />
       <Redirect from="/videos/create" to="/create" />
       <Route path="/videos/:id" component={VideoDisplay} />
       <Route path="/videos/:id/upload" component={VideoUpload} />
